requirements.txt
# Deep Learning & Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# VLM Integration
ollama>=0.1.0

# Data Processing & Output
pyyaml>=6.0
tqdm>=4.65.0

# Image Processing
scikit-image>=0.21.0
scipy>=1.10.0
